pyproject.toml
[project]
name = "yt-transcriber"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "openai-whisper>=20250625",
    "yt-dlp>=2026.1.29",
]
